+++ src/index.ts
@@ -2,6 +2,7 @@ export { Engine } from './engine';
 export * from './types';
 export { VERSION, PROGRAM_ID, MARKET_DEPTH } from './constants';
 export { getSpotPriceStep, getPerpPriceStep } from './engine/utils';
+export { FillEstimateEngine, OrderSide } from './engine/fill-estimate';
 export * from './structure_models';
 export { AccountType } from './types';
 export * from './logs_models';