+++ src/types/index.ts
@@ -18,6 +18,7 @@ export {
   PerpStatisticsResetArgs,
   PerpChangeLeverageArgs,
   PerpForcedCloseArgs,
+  EstimateArgs,
   GetInstrIdArgs,
   GetSpotContextArgs,
   getInstrAccountByTagArgs,
@@ -63,6 +64,7 @@ export {
   ClientCommunityData,
   ClientRefProgramData,
   GetClientDataResponse,
+  EstimateResult,
 } from './responses';
 
 export { LogMessage } from './log-message';
@@ -88,6 +90,7 @@ export {
   PerpChangeLeverageArgsSchema,
   PerpStatisticsResetArgsSchema,
   PerpForcedCloseArgsSchema,
+  EstimateArgsSchema,
   NewInstrumentArgsSchema,
   GetInstrIdArgsSchema,
   GetClientSpotOrdersInfoArgsSchema,